feat(prepaidTier): move dispatchFunds function to dispatcher.js
